refactor: 누락된 barrel export 적용
