+++ components/molecules/index.ts
@@ -1,6 +1,7 @@
 export * from './bottom-sheet';
 export * from './dialog';
 export * from './header-bar';
+export * from './infinite-scroller';
 export * from './page-modal';
 export * from './profile-list-item';
 export * from './record-mark';
