+++ hooks/index.ts
@@ -2,3 +2,4 @@ export * from './use-bottom-sheet';
 export * from './use-calendar-data';
 export * from './use-drag-scroll';
 export * from './use-intersection-observer';
+export * from './use-timeline';
